feat: MarketListCellViewModel 타입 구현
* Cell에 대한 ViewModel 구현
  - Cell에 들어갈 요소들을 ProductData라는 하나의 타입으로 정리
  - Cell의 View 상태를 State enum으로 구현
  - Thumbnail이미지를 fetch해오는 메서드, thumbnailTask를 통제하는 메서드,
    Label에 보일 Text Format을 정하는 메서드 구현
  - 문자열을 하나의 Enum NameSpace에 구현
* Text Format을 위해 Int와 String에 Extension 구현
* 모델 타입인 Product에서 가격관련 프로퍼티를 Int 타입으로 변경

+++ OpenMarket/OpenMarket/Models/Product.swift
@@ -13,10 +13,10 @@ struct Product: Decodable {
     let name: String
     let thumbnailURL: String
     let currency: String
-    let price: Double
+    let price: Int
     let description: String?
-    let bargainPrice: Double
-    let discountedPrice: Double
+    let bargainPrice: Int
+    let discountedPrice: Int
     let stock: Int
     let createdDate: String
     let issuedDate: String
